+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Tweak LMR params: tacticals less, quiets more (7.42 +- 4.53)
 * The TT score is now used for RFP if conditions allow (4.11 +- 3.09)
 * Add late move pruning in quiescence (3.51 +- 2.75)
+* Add hindsight extensions (3.31 +- 2.61)
 * Give a higher history bonus for moves that failed high but where the static eval was below alpha (2.66 +- 2.14)
 * Add 4-ply continuation history (2.16 +- 1.73)
 * LMR is now done later in the root node (0.50 +- 1.83)